+++ cmd/api/middleware/init.go
@@ -1,0 +1,7 @@
+package middleware
+
+import "github.com/galaxy-future/BridgX/cmd/api/middleware/validation"
+
+func Init() {
+	validation.RegisterCustom()
+}
