+++ cmd/api/middleware/validation/README
@@ -1,0 +1,11 @@
+# Validation
+## customer validator
+```go
+
+```
+## gin validation 
+```go
+
+
+
+```
