+++ cmd/api/middleware/validation/validator.go
@@ -9,41 +9,39 @@ import (
 	"github.com/go-playground/validator/v10/translations/zh"
 )
 
-const (
-	CharTypeGT3         = "charTypeGT3"
-	CharTypeGT3TransErr = "必须同时包含三项（大写字母、小写字母、数字、 ()`~!@#$%^&*_-+=|{}[]:;'<>,.?/ 中的特殊符号）"
-)
+type Validation struct {
+	validateFunc     func(validator.FieldLevel) bool
+	translateFunc    func(ut ut.Translator, fe validator.FieldError) string
+	translateRegFunc func(ut ut.Translator) error
+	tag              string
+}
+
+type TagValidatorMap map[string]Validation
+
+// appendTagValidation not thread safe and validation with same tag name will be replaced.
+func appendTagValidation(tag string, validation Validation) {
+	if len(tagMap) == 0 {
+		tagMap = make(map[string]Validation)
+	}
+	tagMap[tag] = validation
+}
+
+// appendMultiTagValidation not thread safe.
+func appendMultiTagValidation(validations ...Validation) {
+	if len(validations) == 0 {
+		return
+	}
+	for _, v := range validations {
+		appendTagValidation(v.tag, v)
+	}
+}
 
 var (
 	zhTranslator            = initZhTranslator()
 	defaultTranslateFunc    = func(ut ut.Translator, fe validator.FieldError) string { return "参数有误" }
 	defaultTranslateRegFunc = func(ut ut.Translator) error { return nil }
-)
 
-type Validation struct {
-	validateFunc     func(validator.FieldLevel) bool
-	translateFunc    func(ut ut.Translator, fe validator.FieldError) string
-	translateRegFunc func(ut ut.Translator) error
-}
-
-var (
-	numberMap      = map[byte]struct{}{'1': {}, '2': {}, '3': {}, '4': {}, '5': {}, '6': {}, '7': {}, '8': {}, '9': {}, '0': {}}
-	upperLetterMap = map[byte]struct{}{'A': {}, 'B': {}, 'C': {}, 'D': {}, 'E': {}, 'F': {}, 'G': {}, 'H': {}, 'I': {}, 'J': {},
-		'K': {}, 'L': {}, 'M': {}, 'N': {}, 'O': {}, 'P': {}, 'Q': {}, 'R': {}, 'S': {}, 'T': {}, 'U': {}, 'V': {}, 'W': {}, 'X': {}, 'Y': {}, 'Z': {}}
-	lowerLetterMap = map[byte]struct{}{'a': {}, 'b': {}, 'c': {}, 'd': {}, 'e': {}, 'f': {}, 'g': {}, 'h': {}, 'i': {}, 'j': {}, 'k': {}, 'l': {},
-		'm': {}, 'n': {}, 'o': {}, 'p': {}, 'q': {}, 'r': {}, 's': {}, 't': {}, 'u': {}, 'v': {}, 'w': {}, 'x': {}, 'y': {}, 'z': {}}
-	specialCharMap = map[byte]struct{}{
-		'(': {}, ')': {}, '`': {}, '~': {}, '!': {}, '@': {}, '#': {}, '$': {}, '%': {}, '^': {}, '&': {}, '*': {}, '_': {},
-		'-': {}, '+': {}, '=': {}, '|': {}, '{': {}, '}': {}, '[': {}, ']': {}, ':': {}, ';': {}, '\'': {}, '<': {}, '>': {}, ',': {}, '.': {}, '?': {}, '/': {},
-	}
-
-	tagMap = map[string]Validation{
-		CharTypeGT3: {
-			validateFunc:     validateCharacterTypeGT3,
-			translateFunc:    translateCharacterErr,
-			translateRegFunc: defaultTranslateRegFunc,
-		},
-	}
+	tagMap = make(map[string]Validation)
 )
 
 func initZhTranslator() ut.Translator {
@@ -52,47 +50,19 @@ func initZhTranslator() ut.Translator {
 	return trans
 }
 
-func validateCharacterTypeGT3(fl validator.FieldLevel) bool {
-	field := []byte(fl.Field().String())
-	var numType, upperLetterType, loweLetterType, specialChatType int
-	for _, c := range field {
-		_, ok := numberMap[c]
-		if ok && numType == 0 {
-			numType = 1
-		}
-		_, ok = upperLetterMap[c]
-		if ok && upperLetterType == 0 {
-			upperLetterType = 1
-		}
-		_, ok = lowerLetterMap[c]
-		if ok && loweLetterType == 0 {
-			loweLetterType = 1
-		}
-		_, ok = specialCharMap[c]
-		if ok && specialChatType == 0 {
-			specialChatType = 1
-		}
-	}
-	return numType+upperLetterType+loweLetterType+specialChatType >= 3
-}
-
-func translateCharacterErr(ut ut.Translator, fe validator.FieldError) string {
-	return "必须同时包含三项（大写字母、小写字母、数字、 ()`~!@#$%^&*_-+=|{}[]:;'<>,.?/ 中的特殊符号）"
-}
-
 func RegisterTools(v *validator.Validate) error {
 	err := registerZHTranslator(v)
 	if err != nil {
 		return err
 	}
-	err = registerCustomerValidation(v)
+	err = registerCustomValidation(v)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func registerCustomerValidation(v *validator.Validate) error {
+func registerCustomValidation(v *validator.Validate) error {
 	if v == nil {
 		return errors.New("empty validator")
 	}
@@ -112,6 +82,7 @@ func registerCustomerValidation(v *validator.Validate) error {
 func registerZHTranslator(v *validator.Validate) error {
 	return zh.RegisterDefaultTranslations(v, zhTranslator)
 }
+
 func Translate2Chinese(err error) string {
 	if err == nil {
 		return ""
